+++ pk3DS/Subforms/Gen6/Experimental/OWSEStructs.cs
@@ -7,802 +7,6 @@ using System.Linq;
 
 namespace pk3DS
 {
-    public class RNG
-    {
-        public static uint Forward32(uint seed, int ctr)
-        {
-            for (int i = 0; i < ctr; i++)
-            {
-                seed *= 0x41C64E6D;
-                seed += 0x00006073;
-            }
-            return seed;
-        }
-    }
-
-    #region Pokémon Related Classes
-    public class Trainer
-    {
-        public bool isORAS;
-        public int Format, Class;
-        public bool Item, Moves;
-        public byte BattleType, NumPokemon, AI;
-        public ushort[] Items = new ushort[4];
-        public byte u1, u2, u3;
-        public ushort uORAS;
-        public bool Healer;
-        public byte Money;
-        public ushort Prize;
-        public Pokemon[] Team;
-        public Trainer(byte[] trData, byte[] trPoke, bool ORAS)
-        {
-            using (BinaryReader br = new BinaryReader(new MemoryStream(trData)))
-            {
-                isORAS = ORAS;
-                Format = ORAS ? br.ReadUInt16() : br.ReadByte();
-                Class = ORAS ? br.ReadUInt16() : br.ReadByte();
-                if (ORAS) uORAS = br.ReadUInt16();
-                Item = ((Format >> 1) & 1) == 1;
-                Moves = (Format & 1) == 1;
-                BattleType = br.ReadByte();
-                NumPokemon = br.ReadByte();
-                for (int i = 0; i < 4; i++)
-                    Items[i] = br.ReadUInt16();
-                AI = br.ReadByte();
-                u1 = br.ReadByte();
-                u2 = br.ReadByte();
-                u3 = br.ReadByte();
-                Healer = br.ReadByte() != 0;
-                Money = br.ReadByte();
-                Prize = br.ReadUInt16();
-
-                // Fetch Team
-                Team = new Pokemon[NumPokemon];
-                byte[][] TeamData = new byte[NumPokemon][];
-                int dataLen = trPoke.Length / NumPokemon;
-                for (int i = 0; i < TeamData.Length; i++)
-                    TeamData[i] = trPoke.Skip(i * dataLen).Take(dataLen).ToArray();
-                for (int i = 0; i < NumPokemon; i++)
-                    Team[i] = new Pokemon(TeamData[i], Item, Moves);
-            }
-        }
-        public byte[] Write()
-        {
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                Format = Convert.ToByte(Moves) + (Convert.ToByte(Item) << 1);
-                if (isORAS)
-                { bw.Write((ushort)Format); bw.Write((ushort)Class); bw.Write((ushort)0); }
-                else
-                { bw.Write((byte)Format); bw.Write((byte)Class); }
-
-                bw.Write(BattleType);
-                bw.Write(NumPokemon);
-                bw.Write(Items[0]);
-                bw.Write(Items[1]);
-                bw.Write(Items[2]);
-                bw.Write(Items[3]);
-
-                bw.Write(AI);
-                bw.Write(u1);
-                bw.Write(u2);
-                bw.Write(u3);
-                bw.Write(Convert.ToByte(Healer));
-                bw.Write(Money);
-                bw.Write(Prize);
-
-                return ms.ToArray();
-            }
-        }
-        public byte[] WriteTeam()
-        {
-            return Team.Aggregate(new byte[0], (i, pkm) => i.Concat(pkm.Write(Item, Moves)).ToArray());
-        }
-
-        public class Pokemon
-        {
-            public byte IVs;
-            public byte PID;
-            public ushort Level;
-            public ushort Species;
-            public ushort Form;
-            public int Ability;
-            public int Gender;
-            public int uBit;
-            public ushort Item;
-            public ushort[] Moves = new ushort[4];
-
-            public Pokemon(byte[] data, bool HasItem, bool HasMoves)
-            {
-                using (BinaryReader br = new BinaryReader(new MemoryStream(data)))
-                {
-                    IVs = br.ReadByte();
-                    PID = br.ReadByte();
-                    Level = br.ReadUInt16();
-                    Species = br.ReadUInt16();
-                    Form = br.ReadUInt16();
-
-                    Ability = PID >> 4;
-                    Gender = PID & 3;
-                    uBit = (PID >> 3) & 1;
-
-                    if (HasItem)
-                        Item = br.ReadUInt16();
-                    if (HasMoves)
-                        for (int i = 0; i < 4; i++)
-                            Moves[i] = br.ReadUInt16();
-                }
-            }
-            public byte[] Write(bool HasItem, bool HasMoves)
-            {
-                using (MemoryStream ms = new MemoryStream())
-                using (BinaryWriter bw = new BinaryWriter(ms))
-                {
-                    bw.Write(IVs);
-                    PID = (byte)(((Ability & 0xF) << 4) | ((uBit & 1) << 3) | (Gender & 0x7));
-                    bw.Write(PID);
-                    bw.Write(Level);
-                    bw.Write(Species);
-                    bw.Write(Form);
-
-                    if (HasItem)
-                        bw.Write(Item);
-                    if (HasMoves)
-                        foreach (ushort Move in Moves)
-                            bw.Write(Move);
-                    return ms.ToArray();
-                }
-            }
-        }
-    }
-    public class Move
-    {
-        public byte Type, Quality, Category, Power, Accuracy, PP, Priority, InflictPercent,
-            HitMin, HitMax, TurnMin, TurnMax, CritStage, Flinch, Recoil, Targeting,
-            Stat1, Stat2, Stat3,
-            Stat1Stage, Stat2Stage, Stat3Stage,
-            Stat1Percent, Stat2Percent, Stat3Percent;
-
-        public Heal Healing;
-        public ushort Inflict, Effect;
-        public byte _0xB, _0x1E, _0x1F, _0x20, _0x21;
-        public Move(byte[] data)
-        {
-            Type = data[0];
-            Quality = data[1];
-            Category = data[2];
-            Power = data[3];
-            Accuracy = data[4];
-            PP = data[5];
-            Priority = data[6];
-            HitMin = (byte)(data[7] & 0xF);
-            HitMax = (byte)(data[7] >> 4);
-            Inflict = BitConverter.ToUInt16(data, 0x8);
-            InflictPercent = data[0xA];
-            _0xB = data[0xB];
-            TurnMin = data[0xC];
-            TurnMax = data[0xD];
-            CritStage = data[0xE];
-            Flinch = data[0xF];
-            Effect = BitConverter.ToUInt16(data, 0x10);
-            Recoil = data[0x12];
-            Healing = new Heal(data[0x13]);
-            Targeting = data[0x14];
-            Stat1 = data[0x15];
-            Stat2 = data[0x16];
-            Stat3 = data[0x17];
-            Stat1Stage = data[0x18];
-            Stat2Stage = data[0x19];
-            Stat3Stage = data[0x1A];
-            Stat1Percent = data[0x1B];
-            Stat2Percent = data[0x1C];
-            Stat3Percent = data[0x1D];
-            _0x1E = data[0x1E];
-            _0x1F = data[0x1F];
-            _0x20 = data[0x20];
-            _0x21 = data[0x21];
-        }
-        public byte[] Write()
-        {
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                bw.Write(Type);
-                bw.Write(Quality);
-                bw.Write(Category);
-                bw.Write(Power);
-                bw.Write(Accuracy);
-                bw.Write(PP);
-                bw.Write(Priority);
-                bw.Write((byte)(HitMin | (HitMax << 4)));
-                bw.Write(Inflict);
-                bw.Write(InflictPercent);
-                bw.Write(_0xB);
-                bw.Write(TurnMin);
-                bw.Write(TurnMax);
-                bw.Write(CritStage);
-                bw.Write(Flinch);
-                bw.Write(Effect);
-                bw.Write(Recoil);
-                bw.Write(Healing.Write());
-                bw.Write(Targeting);
-                bw.Write(Stat1);
-                bw.Write(Stat2);
-                bw.Write(Stat3);
-                bw.Write(Stat1Stage);
-                bw.Write(Stat2Stage);
-                bw.Write(Stat3Stage);
-                bw.Write(Stat1Percent);
-                bw.Write(Stat2Percent);
-                bw.Write(Stat3Percent);
-                bw.Write(_0x1E);
-                bw.Write(_0x1F);
-                bw.Write(_0x20);
-                bw.Write(_0x21);
-                return ms.ToArray();
-            }
-        }
-        public class Heal
-        {
-            public byte Val;
-            public bool Full, Half, Quarter, Value;
-            public Heal(byte val)
-            {
-                Val = val;
-                Full = Val == 0xFF;
-                Half = Val == 0xFE;
-                Quarter = Val == 0xFD;
-                Value = Val < 0xFD;
-            }
-            public byte Write()
-            {
-                if (Value)
-                    return Val;
-                if (Full)
-                    return 0xFF;
-                if (Half)
-                    return 0xFE;
-                if (Quarter)
-                    return 0xFD;
-                return Val;
-            }
-        }
-    }
-    public class Item
-    {
-        public ushort Price;
-        public int BuyPrice;
-        public int SellPrice;
-        public byte HeldEffect, HeldArgument, NaturalGiftEffect, FlingEffect, FlingPower, NaturalGiftPower;
-        public byte NaturalGiftType, u8Flags;
-        public byte KeyFlags;
-        public byte UseEffect;
-        public byte _0xB; // Battle Type
-        public byte _0xC; // 0 or 1
-        public byte _0xD; // Classification (0-3 Battle, 4 Balls, 5 Mail)
-        public byte Consumable;
-        public byte SortIndex;
-        public byte CureInflict; // Bitflags
-        public byte FieldEffect; // Revive 1, Sacred Ash 3, Rare Candy 5, EvoStone 8
-        public int BoostATK, 
-            BoostDEF, BoostSPA, 
-            BoostSPD, BoostSPE,
-            BoostACC, BoostCRIT, BoostPP;
-        public ushort FunctionFlags;
-
-        public ushort FieldBoost; // 0x15-0x16
-        public byte EVHP, EVATK, EVDEF, EVSPE, EVSPA, EVSPD;
-        public Heal Healing;
-        public byte PPGain, Friendship1, Friendship2, Friendship3;
-        public byte _0x23, _0x24;
-
-        public Item(byte[] data)
-        {
-            using (BinaryReader br = new BinaryReader(new MemoryStream(data)))
-            {
-                Price = br.ReadUInt16();
-                BuyPrice = Price*10;
-                SellPrice = Price*5;
-
-                HeldEffect = br.ReadByte();
-                HeldArgument = br.ReadByte();
-                NaturalGiftEffect = br.ReadByte();
-                FlingEffect = br.ReadByte();
-                FlingPower = br.ReadByte();
-                NaturalGiftPower = br.ReadByte();
-                NaturalGiftType = br.ReadByte();
-                u8Flags = (byte)(NaturalGiftType >> 5);
-                NaturalGiftType &= 0x1F;
-                KeyFlags = br.ReadByte();
-                UseEffect = br.ReadByte();
-                _0xB = br.ReadByte();
-                _0xC = br.ReadByte();
-                _0xD = br.ReadByte();
-                Consumable = br.ReadByte();
-                SortIndex = br.ReadByte();
-                CureInflict = br.ReadByte();
-
-                FieldEffect = br.ReadByte();
-                BoostATK = FieldEffect >> 4;
-                FieldEffect &= 0xF;
-                BoostDEF = br.ReadByte();
-                BoostSPA = BoostDEF >> 4;
-                BoostDEF &= 0xF;
-                BoostSPD = br.ReadByte();
-                BoostSPE = BoostSPD >> 4;
-                BoostSPD &= 0xF;
-                BoostACC = br.ReadByte();
-                BoostCRIT = (BoostACC >> 4) & 0x3;
-                BoostPP = BoostACC >> 6;
-                BoostACC &= 0xF;
-
-                FunctionFlags = br.ReadUInt16();
-
-                EVHP = br.ReadByte();
-                EVATK = br.ReadByte();
-                EVDEF = br.ReadByte();
-                EVSPE = br.ReadByte();
-                EVSPA = br.ReadByte();
-                EVSPD = br.ReadByte();
-
-                Healing = new Heal(br.ReadByte());
-                PPGain = br.ReadByte();
-                Friendship1 = br.ReadByte();
-                Friendship2 = br.ReadByte();
-                Friendship3 = br.ReadByte();
-                _0x23 = br.ReadByte();
-                _0x24 = br.ReadByte();
-            }
-        }
-        public byte[] Write()
-        {
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                bw.Write(Price);
-                bw.Write(HeldEffect);
-                bw.Write(HeldArgument);
-                bw.Write(NaturalGiftEffect);
-                bw.Write(FlingEffect);
-                bw.Write(FlingPower);
-                bw.Write(NaturalGiftPower);
-                bw.Write((byte)(NaturalGiftType | (u8Flags << 5)));
-                bw.Write(KeyFlags);
-                bw.Write(UseEffect);
-                bw.Write(_0xB);
-                bw.Write(_0xC);
-                bw.Write(_0xD);
-                bw.Write(Consumable);
-                bw.Write(SortIndex);
-                bw.Write(CureInflict);
-
-                bw.Write((byte)(FieldEffect | (BoostATK << 4)));
-                bw.Write((byte)(BoostDEF | (BoostSPA << 4)));
-                bw.Write((byte)(BoostSPD | (BoostSPE << 4)));
-                bw.Write((byte)(BoostACC | (BoostCRIT << 4) | (BoostPP << 6)));
-
-                bw.Write(FunctionFlags);
-
-                bw.Write(EVHP);
-                bw.Write(EVATK);
-                bw.Write(EVDEF);
-                bw.Write(EVSPE);
-                bw.Write(EVSPA);
-                bw.Write(EVSPD);
-                bw.Write(Healing.Write());
-                bw.Write(PPGain);
-                bw.Write(Friendship1);
-                bw.Write(Friendship2);
-                bw.Write(Friendship3);
-
-                bw.Write(_0x23);
-                bw.Write(_0x24);
-
-                return ms.ToArray();
-            }
-        }
-        public class Heal
-        {
-            public byte Val;
-            public bool Full, Half, Quarter, Value;
-            public Heal(byte val)
-            {
-                Val = val;
-                Full = Val == 0xFF;
-                Half = Val == 0xFE;
-                Quarter = Val == 0xFD;
-                Value = Val < 0xFD;
-            }
-            public byte Write()
-            {
-                if (Value)
-                    return Val;
-                if (Full)
-                    return 0xFF;
-                if (Half)
-                    return 0xFE;
-                if (Quarter)
-                    return 0xFD;
-                return Val;
-            }
-        }
-    }
-    public class Maison
-    {
-        public class Trainer
-        {
-            public ushort Class;
-            public ushort Count;
-            public ushort[] Choices;
-
-            public Trainer() { }
-            public Trainer(byte[] data)
-            {
-                Class = BitConverter.ToUInt16(data, 0);
-                Count = BitConverter.ToUInt16(data, 0);
-                Choices = new ushort[Count];
-                for (int i = 0; i < Count; i++)
-                    Choices[i] = BitConverter.ToUInt16(data, 4 + 2 * i);
-            }
-            public byte[] Write()
-            {
-                using (var ms = new MemoryStream())
-                using (var bw = new BinaryWriter(ms))
-                {
-                    bw.Write(Class);
-                    bw.Write(Count);
-                    foreach (ushort Choice in Choices)
-                        bw.Write(Choice);
-                    return ms.ToArray();
-                }
-            }
-        }
-        public class Pokemon
-        {
-            public ushort Species, Item;
-            public byte EVs, Nature;
-            public ushort[] Moves;
-            public bool HP, ATK, DEF, SPE, SPA, SPD;
-
-            private readonly byte _u1;
-            private readonly byte _u2;
-
-            public Pokemon(byte[] data)
-            {
-                Species = BitConverter.ToUInt16(data, 0);
-                Moves = new[]
-                {
-                    BitConverter.ToUInt16(data, 2),
-                    BitConverter.ToUInt16(data, 4),
-                    BitConverter.ToUInt16(data, 6),
-                    BitConverter.ToUInt16(data, 8)
-                };
-                EVs = data[0xA];
-                HP = (EVs >> 0 & 1) == 1;
-                ATK = (EVs >> 1 & 1) == 1;
-                DEF = (EVs >> 2 & 1) == 1;
-                SPE = (EVs >> 3 & 1) == 1;
-                SPA = (EVs >> 4 & 1) == 1;
-                SPD = (EVs >> 5 & 1) == 1;
-                Nature = data[0xB];
-                Item = BitConverter.ToUInt16(data, 0xC);
-                _u1 = data[0xE];
-                _u2 = data[0xF];
-            }
-            public byte[] Write()
-            {
-                using (var ms = new MemoryStream())
-                using (var bw = new BinaryWriter(ms))
-                {
-                    bw.Write(Species);
-                    foreach (ushort Move in Moves)
-                        bw.Write(Move);
-
-                    EVs &= 0xC0;
-                    EVs |= (byte)(HP ? 1 << 0 : 0);
-                    EVs |= (byte)(ATK ? 1 << 1 : 0);
-                    EVs |= (byte)(DEF ? 1 << 1 : 0);
-                    EVs |= (byte)(SPE ? 1 << 1 : 0);
-                    EVs |= (byte)(SPA ? 1 << 1 : 0);
-                    EVs |= (byte)(SPD ? 1 << 1 : 0);
-                    bw.Write(EVs);
-
-                    bw.Write(Nature);
-                    bw.Write(Item);
-                    bw.Write(_u1);
-                    bw.Write(_u2);
-                    return ms.ToArray();
-                }
-            }
-        }
-    }
-    public class Learnset
-    {
-        public int Count;
-        public short[] Moves;
-        public short[] Levels;
-        public Learnset(byte[] data)
-        {
-            if (data.Length < 4 || data.Length % 4 != 0) return; // Detect invalid files, weakly.
-            Count = data.Length / 4 - 1;
-            Moves = new short[Count];
-            Levels = new short[Count];
-            using (BinaryReader br = new BinaryReader(new MemoryStream(data)))
-            {
-                for (int i = 0; i < Count; i++)
-                {
-                    Moves[i] = br.ReadInt16();
-                    Levels[i] = br.ReadInt16();
-                }
-            }
-        }
-        public byte[] Write()
-        {
-            Count = (ushort)Moves.Length;
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                for (int i = 0; i < Count; i++)
-                {
-                    bw.Write(Moves[i]);
-                    bw.Write(Levels[i]);
-                }
-                bw.Write(-1);
-                return ms.ToArray();
-            }
-        }
-    }
-    public class EggMoves
-    {
-        public ushort Count;
-        public ushort[] Moves;
-
-        public EggMoves(byte[] data)
-        {
-            if (data.Length < 2 || data.Length % 2 != 0) return; // Detect invalid files, weakly.
-            using (BinaryReader br = new BinaryReader(new MemoryStream(data)))
-            {
-                Count = br.ReadUInt16();
-                Moves = new ushort[Count];
-                for (int i = 0; i < Count; i++)
-                    Moves[i] = br.ReadUInt16();
-            }
-        }
-        public byte[] Write()
-        {
-            Count = (ushort)Moves.Length;
-            if (Count == 0) return new byte[0];
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                bw.Write(Count);
-                for (int i = 0; i < Count; i++)
-                    bw.Write(Moves[i]);
-
-                return ms.ToArray();
-            }
-        }
-    }
-    public class Evolutions
-    {
-        public ushort[] Method, Criteria, Species;
-        public Evolutions(byte[] data)
-        {
-            if (data.Length < 0x30 || data.Length % 6 != 0) return;
-            Method = new ushort[data.Length / 6];
-            Criteria = new ushort[data.Length / 6];
-            Species = new ushort[data.Length / 6];
-            using (BinaryReader br = new BinaryReader(new MemoryStream(data)))
-            for (int i = 0; i < Species.Length; i++)
-            {
-                Method[i] = br.ReadUInt16();
-                Criteria[i] = br.ReadUInt16();
-                Species[i] = br.ReadUInt16();
-            }
-        }
-        public byte[] Write()
-        {
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                for (int i = 0; i < Species.Length; i++)
-                {
-                    bw.Write(Method[i]);
-                    bw.Write(Criteria[i]);
-                    bw.Write(Species[i]);
-                }
-                return ms.ToArray();
-            }
-        }
-    }
-    public class Gift
-    {
-        // All
-        public byte[] Data;
-        public bool ORAS;
-
-        public ushort Species;
-        public ushort u2;
-        public byte Form;
-        public byte Level;
-        public sbyte Ability;
-        public sbyte Nature;
-        public byte Shiny, u9, uA, uB;
-        public int HeldItem;
-        public sbyte Gender;
-        // ORAS
-        public byte u11;
-        public short MetLocation;
-        public ushort Move;
-        // All
-        public sbyte[] IVs = new sbyte[6];
-        // ORAS
-        public byte[] ContestStats;
-        public byte u22;
-        // All
-        public byte uLast;
-
-        public Gift(byte[] data, bool oras)
-        {
-            Data = data;
-            ORAS = oras;
-            using (BinaryReader br = new BinaryReader(new MemoryStream(Data)))
-            {
-                Species = br.ReadUInt16();
-                u2 = br.ReadUInt16();
-                Form = br.ReadByte();
-                Level = br.ReadByte();
-                Shiny = br.ReadByte();
-                Ability = br.ReadSByte();
-                Nature = br.ReadSByte();
-                u9 = br.ReadByte();
-                uA = br.ReadByte();
-                uB = br.ReadByte();
-                HeldItem = br.ReadInt32();
-                Gender = br.ReadSByte();
-
-                if (ORAS)
-                {
-                    u11 = br.ReadByte();
-                    MetLocation = br.ReadInt16();
-                    Move = br.ReadUInt16();
-                }
-
-                for (int i = 0; i < 6; i++)
-                    IVs[i] = br.ReadSByte();
-
-                if (ORAS)
-                {
-                    ContestStats = br.ReadBytes(6);
-                    u22 = br.ReadByte();
-                }
-
-                uLast = br.ReadByte();
-            }
-        }
-        public byte[] Write()
-        {
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                bw.Write(Species);
-                bw.Write(u2);
-                bw.Write(Form);
-                bw.Write(Level);
-                bw.Write(Shiny);
-                bw.Write(Ability);
-                bw.Write(Nature);
-                bw.Write(u9);
-                bw.Write(uA);
-                bw.Write(uB);
-                bw.Write(HeldItem);
-                bw.Write(Gender);
-
-                if (ORAS)
-                {
-                    bw.Write(u11);
-                    bw.Write(MetLocation);
-                    bw.Write(Move);
-                }
-
-                for (int i = 0; i < 6; i++)
-                    bw.Write(IVs[i]);
-
-                if (ORAS)
-                {
-                    bw.Write(ContestStats);
-                    bw.Write(u22);
-                }
-
-                bw.Write(uLast);
-
-                return ms.ToArray();
-            }
-        }
-    }
-    public class EncounterStatic
-    {
-        // All
-        public readonly byte[] Data;
-
-        public ushort Species
-        {
-            get { return BitConverter.ToUInt16(Data, 0x0); }
-            set { BitConverter.GetBytes(value).CopyTo(Data, 0x0); }
-        }
-        public byte Form { get { return Data[0x2]; } set { Data[0x2] = value; } }
-        public byte Level { get { return Data[0x3]; } set { Data[0x3] = value; } }
-        public int HeldItem
-        {
-            get
-            {
-                int item = BitConverter.ToInt16(Data, 0x4);
-                if (item < 0) item = 0;
-                return item;
-            }
-            set
-            {
-                if (value == 0) value = -1;
-                BitConverter.GetBytes((short)value).CopyTo(Data, 0x4);
-            }
-        }
-        
-        public bool ShinyLock { get { return (Data[0x6] & 2) >> 1 == 1; } set { Data[0x6] = (byte)(Data[0x6] & ~2 | (value ? 2 : 0)); } }
-        public int Gender { get { return (Data[0x6] & 0xC) >> 2; } set { Data[0x6] = (byte)(Data[0x6] & ~0xC | ((value & 3) << 2)); } }
-        public int Ability { get { return (Data[0x6] & 0x70) >> 4; } set { Data[0x6] = (byte)(Data[0x6] & ~0x70 | ((value & 7) << 4)); } }
-
-        public bool IV3 { get { return (Data[0x7] & 1) >> 0 == 1; } set { Data[0x7] = (byte)(Data[0x7] & ~1 | (value ? 1 : 0)); } }
-        public bool IV3_1 { get { return (Data[0x7] & 2) >> 1 == 1; } set { Data[0x7] = (byte)(Data[0x7] & ~2 | (value ? 2 : 0)); } }
-
-        public EncounterStatic(byte[] data)
-        {
-            Data = (byte[])data.Clone();
-        }
-        public byte[] Write()
-        {
-            return (byte[])Data.Clone();
-        }
-    }
-    public class MegaEvolutions
-    {
-        public ushort[] Form, Method, Argument, u6;
-        public MegaEvolutions(byte[] data)
-        {
-            if (data.Length < 0x10 || data.Length % 8 != 0) return;
-            Form = new ushort[data.Length / 8];
-            Method = new ushort[data.Length / 8];
-            Argument = new ushort[data.Length / 8];
-            u6 = new ushort[data.Length / 8];
-            using (BinaryReader br = new BinaryReader(new MemoryStream(data)))
-            for (int i = 0; i < Form.Length; i++)
-            {
-                Form[i] = br.ReadUInt16();
-                Method[i] = br.ReadUInt16();
-                Argument[i] = br.ReadUInt16();
-                u6[i] = br.ReadUInt16();
-            }
-        }
-        public byte[] Write()
-        {
-            using (MemoryStream ms = new MemoryStream())
-            using (BinaryWriter bw = new BinaryWriter(ms))
-            {
-                for (int i = 0; i < Form.Length; i++)
-                {
-                    if (Method[i] == 0)
-                    { Form[i] = Argument[i] = 0; } // No method to evolve, clear information.
-                    bw.Write(Form[i]);
-                    bw.Write(Method[i]);
-                    bw.Write(Argument[i]);
-                    bw.Write(u6[i]);
-                }
-                return ms.ToArray();
-            }
-        }
-    }
-    #endregion
-
     #region Game Related Classes
     public class MapMatrix
     {
@@ -924,7 +128,7 @@ namespace pk3DS
                     int Y = i / 40;
                     uint colorValue = Tiles[i] == 0x01000021
                         ? 0xFF000000
-                        : RNG.Forward32(Tiles[i], ColorShift) | 0xFF000000;
+                        : LCRNG32.Advance(Tiles[i], ColorShift) | 0xFF000000;
 
                     byte[] pixel = BitConverter.GetBytes(colorValue);
                     for (int x = 0; x < s * s; x++)
